Fixes #120 Browsing through asynchronous turns
When mouse goes over the target in the event summary of an asynchronous
turn a drop down shows all the turns that are related to the current.
When clicking a previous turn in the event summary, the event summary
changes to display the clicked turn's information.
Changed kodemo to have a chain of asynchronous calls.